Fixed build scripts - Moved src under /logcat - Updated .api files - Updated Gradle plugin to 7.0.0 - Removed versionName and versionCode from android block in build.gradle.kts - Re-added the publishing and KtLint checks - Fixed lint errors - Removed jcenter repo
