+++ gradle.properties
@@ -2,6 +2,7 @@ kotlin.code.style=official
 #kotlin.mpp.enableGranularSourceSetsMetadata=true
 #kotlin.native.enableDependencyPropagation=false
 kotlin.js.generate.executable.default=false
+org.gradle.jvmargs='-Dfile.encoding=UTF-8'
 android.useAndroidX=true
 
 # Required to publish to Nexus (see https://github.com/gradle/gradle/issues/11308)
