+++ logcat/api/logcat.api
@@ -1,18 +1,16 @@
-public final class logcat/AndroidLogcatLogger : logcat/LogcatLogger {
-	public static final field Companion Llogcat/AndroidLogcatLogger$Companion;
-	public fun <init> ()V
-	public fun <init> (Llogcat/LogPriority;)V
-	public synthetic fun <init> (Llogcat/LogPriority;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun isLoggable (Llogcat/LogPriority;)Z
-	public fun log (Llogcat/LogPriority;Ljava/lang/String;Ljava/lang/String;)V
+public final class logcat/DefaultOuterClassSimpleNameInternalOnlyDoNotUseKThxByeKt {
+	public static final fun defaultOuterClassSimpleNameInternalOnlyDoNotUseKThxBye (Ljava/lang/Object;)Ljava/lang/String;
 }
 
-public final class logcat/AndroidLogcatLogger$Companion {
-	public final fun installOnDebuggableApp (Landroid/app/Application;Llogcat/LogPriority;)V
-	public static synthetic fun installOnDebuggableApp$default (Llogcat/AndroidLogcatLogger$Companion;Landroid/app/Application;Llogcat/LogPriority;ILjava/lang/Object;)V
+public final class logcat/DefaultThrowableAsLogKt {
+	public static final fun defaultThrowableAsLog (Ljava/lang/Throwable;)Ljava/lang/String;
 }
 
-public final class logcat/AndroidLogcatLoggerKt {
+public final class logcat/ExtLogcatLoggerKt {
+	public static final fun getLogger (Llogcat/LogcatLogger$Companion;)Llogcat/LogcatLogger;
+	public static final fun install (Llogcat/LogcatLogger$Companion;Llogcat/LogcatLogger;)V
+	public static final fun isInstalled (Llogcat/LogcatLogger$Companion;)Z
+	public static final fun uninstall (Llogcat/LogcatLogger$Companion;)V
 }
 
 public final class logcat/LogPriority : java/lang/Enum {
@@ -32,7 +30,6 @@ public final class logcat/LogcatKt {
 	public static final fun logcat (Ljava/lang/String;Llogcat/LogPriority;Lkotlin/jvm/functions/Function0;)V
 	public static synthetic fun logcat$default (Ljava/lang/Object;Llogcat/LogPriority;Ljava/lang/String;Lkotlin/jvm/functions/Function0;ILjava/lang/Object;)V
 	public static synthetic fun logcat$default (Ljava/lang/String;Llogcat/LogPriority;Lkotlin/jvm/functions/Function0;ILjava/lang/Object;)V
-	public static final fun outerClassSimpleNameInternalOnlyDoNotUseKThxBye (Ljava/lang/Object;)Ljava/lang/String;
 }
 
 public abstract interface class logcat/LogcatLogger {
@@ -42,14 +39,21 @@ public abstract interface class logcat/LogcatLogger {
 }
 
 public final class logcat/LogcatLogger$Companion {
-	public final fun getLogger ()Llogcat/LogcatLogger;
-	public final fun install (Llogcat/LogcatLogger;)V
-	public final fun isInstalled ()Z
-	public final fun uninstall ()V
 }
 
 public final class logcat/LogcatLogger$DefaultImpls {
 	public static fun isLoggable (Llogcat/LogcatLogger;Llogcat/LogPriority;)Z
+}
+
+public final class logcat/NoLog : logcat/LogcatLogger {
+	public static final field INSTANCE Llogcat/NoLog;
+	public fun isLoggable (Llogcat/LogPriority;)Z
+	public fun log (Llogcat/LogPriority;Ljava/lang/String;Ljava/lang/String;)Ljava/lang/Void;
+	public synthetic fun log (Llogcat/LogPriority;Ljava/lang/String;Ljava/lang/String;)V
+}
+
+public final class logcat/OuterClassSimpleNameInternalOnlyDoNotUseKThxByeKt {
+	public static final fun outerClassSimpleNameInternalOnlyDoNotUseKThxBye (Ljava/lang/Object;)Ljava/lang/String;
 }
 
 public final class logcat/PrintLogger : logcat/LogcatLogger {
