+++ logcat/gradle.properties
@@ -1,3 +1,25 @@
-POM_ARTIFACT_ID=logcat
-POM_NAME=Logcat
-POM_PACKAGING=aar
+kotlin.code.style=official
+#kotlin.mpp.enableGranularSourceSetsMetadata=true
+#kotlin.native.enableDependencyPropagation=false
+kotlin.js.generate.executable.default=false
+android.useAndroidX=true
+
+# Required to publish to Nexus (see https://github.com/gradle/gradle/issues/11308)
+systemProp.org.gradle.internal.publish.checksums.insecure=true
+
+GROUP=com.squareup.logcat
+VERSION_NAME=1.0-SNAPSHOT
+
+POM_DESCRIPTION=A tiny Kotlin API for cheap logging on top of Android's normal Log class.
+
+POM_URL=https://github.com/square/logcat/
+POM_SCM_URL=https://github.com/square/logcat/
+POM_SCM_CONNECTION=scm:git:git://github.com/square/logcat.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/square/logcat.git
+
+POM_LICENCE_NAME=The Apache Software License, Version 2.0
+POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
+POM_LICENCE_DIST=repo
+
+POM_DEVELOPER_ID=square
+POM_DEVELOPER_NAME=Square, Inc.
