+++ logcat/src/androidMain/AndroidManifest.xml
