+++ logcat/src/androidMain/kotlin/logcat/outerClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
