+++ logcat/src/commonMain/kotlin/logcat/LogPriority.kt
