+++ logcat/src/commonMain/kotlin/logcat/defaultOuterClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
