+++ logcat/src/commonMain/kotlin/logcat/defaultThrowableAsLog.kt
