+++ logcat/src/commonMain/kotlin/logcat/outerClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
@@ -1,0 +1,4 @@
+package logcat
+
+@PublishedApi
+internal expect fun Any.outerClassSimpleNameInternalOnlyDoNotUseKThxBye(): String
