+++ logcat/src/commonTest/kotlin/logcat/ITestLogcatLogger.kt
@@ -4,6 +4,6 @@ package logcat
  * Extension of the LogcatLogger interface provided in the logcat library that
  * provides access to `latestLog` for inspection during tests.
  */
-interface ITestLogcatLogger: LogcatLogger {
+interface ITestLogcatLogger : LogcatLogger {
   var latestLog: Log?
 }
