+++ logcat/src/commonTest/kotlin/logcat/TestLogcatLogger.kt
@@ -3,7 +3,8 @@ package logcat
 /**
  * Standard implementation of a LogcatLogger that can be used on Android/JVM/JS
  */
-class TestLogcatLogger(private val isLoggable: (LogPriority) -> Boolean = { true }) : ITestLogcatLogger {
+class TestLogcatLogger(private val isLoggable: (LogPriority) -> Boolean = { true }) :
+  ITestLogcatLogger {
   override fun isLoggable(priority: LogPriority): Boolean = isLoggable.invoke(priority)
 
   override var latestLog: Log? = null
