+++ logcat/src/jsMain/kotlin/logcat/outerClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
