+++ logcat/src/jvmMain/kotlin/logcat/outerClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
