+++ logcat/src/jvmTest/kotlin/logcat/platformTestLogger.kt
@@ -1,0 +1,5 @@
+package logcat
+
+actual fun platformTestLogger(isLoggable: (LogPriority) -> Boolean): ITestLogcatLogger {
+  return TestLogcatLogger(isLoggable = isLoggable)
+}
