+++ logcat/src/nativeMain/kotlin/logcat/outerClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
