+++ logcat/src/nativeTest/kotlin/logcat/NativeLogcatTest.kt
@@ -1,12 +1,13 @@
 package logcat
 
 import logcat.LogPriority.INFO
+import kotlin.native.concurrent.AtomicReference
 import kotlin.test.AfterTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
-class JsLogcatTest {
+class NativeLogcatTest {
 
   @AfterTest
   fun tearDown() {
@@ -14,28 +15,27 @@ class JsLogcatTest {
   }
 
   @Test fun logcat_passes_priority_to_isLoggable_check() {
-    var receivedPriority: LogPriority? = null
-    TestLogcatLogger(isLoggable = { receivedPriority = it; true })
+    val receivedPriority: AtomicReference<LogPriority?> = AtomicReference(null)
+    NativeTestLogcatLogger(isLoggable = { receivedPriority.value = it; true })
       .apply { LogcatLogger.install(this); latestLog = null }
 
     logcat(INFO) { "Hi" }
 
-    assertEquals(INFO, receivedPriority)
+    assertEquals(INFO, receivedPriority.value)
   }
 
   @Test fun Throwable_asLogMessage_has_stacktrace_logged() {
-    val logger = TestLogcatLogger().apply { LogcatLogger.install(this); latestLog = null }
+    val logger = NativeTestLogcatLogger().apply { LogcatLogger.install(this); latestLog = null }
     val exception = RuntimeException("damn")
 
     logcat { exception.asLog() }
 
     val stackTraceStr = """
-      |RuntimeException: damn
-      |    at JsLogcatTest.Throwable_asLogMessage_has_stacktrace_logged
+      |kotlin.RuntimeException: damn
+      |    at 0
       """.trimMargin()
     assertTrue(
       logger.latestLog!!.message.contains(stackTraceStr)
     )
   }
-
 }
