+++ settings.gradle.kts
@@ -1,7 +1,6 @@
 pluginManagement {
   repositories {
     google()
-    jcenter()
     gradlePluginPortal()
     mavenCentral()
   }
@@ -15,3 +14,7 @@ pluginManagement {
 }
 rootProject.name = "logcat"
 
+include(
+  ":logcat",
+  ":sample"
+)
