--- src/commonMain/kotlin/logcat/outerClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
@@ -1,4 +1,0 @@
-package logcat
-
-@PublishedApi
-internal expect fun Any.outerClassSimpleNameInternalOnlyDoNotUseKThxBye(): String
