--- src/nativeTest/kotlin/logcat/platformTestLogger.kt
@@ -1,3 +1,0 @@
-package logcat
-
-actual fun platformTestLogger(isLoggable: (LogPriority) -> Boolean): ITestLogcatLogger = NativeTestLogcatLogger(isLoggable = isLoggable)
